requirements.txt
streamlit
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
numpy
scipy
scikit-learn
matplotlib
pandas
tensorflow-cpu
pandas_datareader